fix: tests && border
