+++ 8vim/src/main/kotlin/inc/flide/vim8/ime/ui/ImeLayout.kt
@@ -2,12 +2,13 @@ package inc.flide.vim8.ime.ui
 
 import androidx.compose.runtime.Composable
 import inc.flide.vim8.ime.theme.ImeTheme
+import inc.flide.vim8.ime.ui.floating.compose.Layout
 
 @Composable
 fun ImeLayout(isFloating: Boolean, content: @Composable () -> Unit) {
     ImeTheme {
         if (isFloating) {
-            FloatingImeLayout {
+            Layout {
                 content()
             }
         } else {
