+++ 8vim/src/main/kotlin/inc/flide/vim8/ime/ui/floating/SizeExtension.kt
@@ -1,4 +1,4 @@
-package inc.flide.vim8.ime.ui
+package inc.flide.vim8.ime.ui.floating
 
 import androidx.compose.ui.geometry.Offset
 import androidx.compose.ui.geometry.Size
@@ -18,14 +18,20 @@ internal fun Size.isWithinWidth(screenSize: Size) = width > screenSize.minWidth 
 internal inline val Size.minHeight: Float
     get() = min(height, width) * 0.6f
 internal inline val Size.maxHeight: Float
-    get() = min(height, width) * 0.7f
+    get() = min(height, width) * 0.8f
 internal inline val Size.minWidth: Float
     get() = minHeight
 internal inline val Size.maxWidth: Float
     get() = maxHeight
 
+fun Size.coerceIn(size: Size): Size = Size(
+    width = width.coerceIn(size.minWidth, size.maxWidth),
+    height = height.coerceIn(size.minHeight, size.maxHeight)
+)
+
 fun IntOffset.coerceIn(size: IntSize, screenSize: IntSize): IntOffset =
     toOffset().coerceIn(size.toSize(), screenSize.toSize()).toIntOffset()
+
 fun Offset.coerceIn(size: Size, screenSize: Size): Offset {
     val (minHeight, maxHeight) = if (screenSize.height > size.height) {
         screenSize.height to size.height
